- implement requested refactor on previous PR review
